backend.challenge/indexer/requirements.txt
kafka-python==2.0.2
pymongo==3.5.1